move threshold_li benchmark to benchmark_filters.py
location consistent with other threshold benchmarks (Otsu, Sauvola)
